+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Enable combat tracks without having to import the camping sheet first
+
 ## [0.9.0] - 2023-10-14
 
 ### Added