+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Enable combat tracks without having to import the camping sheet first
+* Fix Sharpshooter scaling after level 8
+* Fix Deploy Army skills
 
 ## [0.9.0] - 2023-10-14